agent/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
langchain>=0.2.16
langchain-openai>=0.1.17
langchain-community>=0.2.16
requests>=2.31.0
#tavily-python>=0.3.3
langchain-tavily
